requirements.txt
boto3==1.34.64
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
scikit_learn==1.3.0
spacy==3.7.4
tensorflow==2.15.0
tensorflow_macos==2.15.0
streamlit==1.30.0
